silence 64-bit warnings.
git-svn-id: https://dev.csoft.net/agar/trunk@5730 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ gui/label.c
@@ -347,31 +347,36 @@ AG_LabelDrawPolled(AG_Label *label)
 					case 'd':
 					case 'i':
 						snprintf(s2, sizeof(s2),
-						    "%lld", LABEL_ARG(Sint64));
+						    "%lld", (long long)
+						    LABEL_ARG(Sint64));
 						strlcat(s, s2, sizeof(s));
 						ri++;
 						break;
 					case 'o':
 						snprintf(s2, sizeof(s2),
-						    "%llo", LABEL_ARG(Uint64));
+						    "%llo", (unsigned long long)
+						    LABEL_ARG(Uint64));
 						strlcat(s, s2, sizeof(s));
 						ri++;
 						break;
 					case 'u':
 						snprintf(s2, sizeof(s2),
-						    "%llu", LABEL_ARG(Uint64));
+						    "%llu", (unsigned long long)
+						    LABEL_ARG(Uint64));
 						strlcat(s, s2, sizeof(s));
 						ri++;
 						break;
 					case 'x':
 						snprintf(s2, sizeof(s2),
-						    "%llx", LABEL_ARG(Uint64));
+						    "%llx", (unsigned long long)
+						    LABEL_ARG(Uint64));
 						strlcat(s, s2, sizeof(s));
 						ri++;
 						break;
 					case 'X':
 						snprintf(s2, sizeof(s2),
-						    "%llX", LABEL_ARG(Uint64));
+						    "%llX", (unsigned long long)
+						    LABEL_ARG(Uint64));
 						strlcat(s, s2, sizeof(s));
 						ri++;
 						break;
